feature: 회원가입 및 로그인 API 기능 구현 완료
- 프론트측에서 요청받은 블로그 고유ID, 사용자명(닉네임), 이메일 중복확인 처리를 AOP 관심사 분리(DuplicateCheck)를 통해 구현
- 회원가입 요청 시 이메일 인증코드를 발급하고, 이메일 인증코드를 입력받아 인증하는 방식으로 회원가입 처리
- 중복확인 요청, 회원가입 요청, 이메일 인증코드 발급, 인증코드 재발급, 로그인 요청 모두 1분에 3번만 가능하도록(RateLimit) 적용 -> 이것 또한 AOP를 이용해 구현

- OAuth2 사용자
 - OAuth2 신규 사용자, 기존 사용자 구분해서 처리
 - OAuth2 신규 사용자 또한 일반 폼 로그인 사용자와 동일한 방식으로 적용하기 위해 추가적으로 사용자명(닉네임), 블로그 고유ID를 등록할 수 있도록 구현
 - OAuth2 로그인 -> 인증 성공 -> 추가 정보 입력 -> 추가 정보 입력 성공 -> 최종적으로 사용자 정보 DB에 저장 및 JWT 토큰 발급.
 - OAuth2 기존 사용자는 전과 동일한 방식으로 유지

- 로그인 요청시 리멤버미
 - 리멤버미가 체크되었다면 리프레시 토큰을 2주로 설정하고, 리멤버미가 체크되지 않았다면 리프레시 토큰을 1일로 설정
 - 액세스 토큰은 리멤버미 체크 유무와 상관없이 1시간 유지
 - 신규 사용자가 추가 정보 입력을 완료하면 리멤버미가 체크되었는지 여부에 따라 리프레시 토큰 만료 시간을 설정해야 하는데, 이때 리멤버미 체크 여부 정보를 유지할 필요가 있음.
 - 이를 위해 OAuth2AuthorizationRequestRedirectFilter앞에 RememberMeAuthenticationFilter를 추가하여 쿠키와 Redis를 이용해 리멤버미 체크 여부를 유지하도록 구현.

- 회원가입 요청시 오는 DTO에 @Valid 추가.
- 이유는 프론트 측에서 유효성 검사를 하긴 했지만 악의적인 사용자가 직접 요청을 보낼 수 있기 때문에 최종적으로는 보안을 위해 서버 측에서 처리하는게 맞다고 판단.

- 리프레시 토큰 관리 방식은 DB가 아닌 Redis에 저장. 즉 기존 RefreshToken을 DB에 저장하는 방식 -> Redis를 이용해 저장하는 방식으로 변경
- 이유는 리프레시 토큰을 검증 또는 업데이트할때 DB에 접근하는 것보다 Redis에 접근하는 것이 더 빠른 측면도 있고, 관리하는 측면에서도 Redis는 직접 삭제할 필요 없이 자동 만료 기능을 제공하기 때문에 더 효율적이라고 판단.

+++ blog-springboot-jpa/src/main/java/com/yhs/blog/springboot/jpa/domain/user/controller/UserApiController.java
@@ -1,12 +1,12 @@
 package com.yhs.blog.springboot.jpa.domain.user.controller;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
+import com.yhs.blog.springboot.jpa.aop.ratelimit.RateLimit;
 import com.yhs.blog.springboot.jpa.common.response.ApiResponse;
 import com.yhs.blog.springboot.jpa.common.response.ErrorResponse;
 import com.yhs.blog.springboot.jpa.common.response.SuccessResponse;
 import com.yhs.blog.springboot.jpa.common.util.cookie.CookieUtil;
 import com.yhs.blog.springboot.jpa.domain.token.jwt.provider.TokenProvider;
-import com.yhs.blog.springboot.jpa.domain.token.jwt.service.RefreshTokenService;
 import com.yhs.blog.springboot.jpa.domain.token.jwt.service.TokenManagementService;
 import com.yhs.blog.springboot.jpa.domain.user.dto.request.LoginRequest;
 import com.yhs.blog.springboot.jpa.domain.user.dto.request.SignUpUserRequest;
@@ -28,6 +28,7 @@ import jakarta.servlet.http.HttpServletResponse;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
+import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -36,9 +37,12 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.*;
 
 import java.io.IOException;
+import java.util.concurrent.TimeUnit;
+
 @Log4j2
 @RestController
 @RequiredArgsConstructor
@@ -47,33 +51,31 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
     private final UserService userService;
     private final AuthenticationManager authenticationManager;
     private final TokenProvider tokenProvider;
-    private final RefreshTokenService refreshTokenService;
     private final TokenManagementService tokenManagementService;
     private final EmailService emailService;
+    private final RedisTemplate<String, String> redisTemplate;
 
-    // 일단 하나씩 해보면서 잘 되는지 회원가입부터 해봐야지
+    // 회원가입시 이메일 인증코드 전송, 인증코드 재전송 부분 공통 처리
     @PostMapping("/api/users/signup")
     public ResponseEntity<ApiResponse> signup(@RequestBody @Valid SignUpUserRequest signUpUserRequest) throws JsonProcessingException {
 
-        log.info("signUpUserRequest: " + signUpUserRequest);
-
-        RateLimitResponse result = emailService.processSignUp(signUpUserRequest);
+        RateLimitResponse result = emailService.processEmailVerification(signUpUserRequest);
         if (result.isSuccess()) {
             return ResponseEntity.status(result.getStatusCode())
                     .body(new SuccessResponse<>(result.getData(), result.getMessage()));
         }
 
-
         return ResponseEntity.status(result.getStatusCode())
                 .body(new ErrorResponse(result.getMessage(), result.getStatusCode()));
 
     }
-
 
     @PostMapping("/api/users/verify-email")
     public ResponseEntity<ApiResponse> verifyEmail(@RequestBody @Valid VerifyEmailRequest verifyEmailRequest) {
 
-        RateLimitResponse result =  emailService.completeVerification(verifyEmailRequest);
+        RateLimitResponse result = emailService.completeVerification(verifyEmailRequest);
+
+        log.info("result: " + result);
 
         if (result.isSuccess()) {
             return ResponseEntity.status(result.getStatusCode())
@@ -84,8 +86,10 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
                 .body(new ErrorResponse(result.getMessage(), result.getStatusCode()));
     }
 
+    @RateLimit(key = "Login")
     @PostMapping("/api/users/login")
-    public ResponseEntity<String> login(@RequestBody LoginRequest loginRequest,
+    @Transactional // saveRefreshToken DB작업 있어서 트랜잭션 추가해야함
+    public ResponseEntity<SuccessResponse<Void>> login(@RequestBody @Valid LoginRequest loginRequest,
                                         HttpServletRequest request,
                                         HttpServletResponse response) throws ServletException, IOException {
 
@@ -98,16 +102,24 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
 
         HttpHeaders headers = new HttpHeaders();
 
+        String refreshToken;
 
-        // 리프레시 토큰 생성
-        String refreshToken = tokenProvider.generateToken(user,
-                TokenManagementService.REFRESH_TOKEN_DURATION);
+        if (loginRequest.getRememberMe()) {
+            refreshToken = tokenProvider.generateToken(user,
+                    TokenManagementService.REMEMBER_ME_REFRESH_TOKEN_DURATION);
+            redisTemplate.opsForValue().set(TokenManagementService.RT_PREFIX + user.getEmail(), refreshToken,
+                    TokenManagementService.REMEMBER_ME_REFRESH_TOKEN_TTL, TimeUnit.SECONDS);
 
-        // 리프레시 토큰을 userId와 함께 DB에 저장
-        tokenManagementService.saveRefreshToken(user.getId(), refreshToken);
+        } else {
+            refreshToken = tokenProvider.generateToken(user, TokenManagementService.REFRESH_TOKEN_DURATION);
+
+            redisTemplate.opsForValue().set(TokenManagementService.RT_PREFIX + user.getEmail(), refreshToken,
+                    TokenManagementService.REFRESH_TOKEN_TTL, TimeUnit.SECONDS);
+
+        }
 
         // 생성된 리프레시 토큰을 클라이언트측 쿠키에 저장 -> 클라이언트에서 액세스 토큰이 만료되면 재발급 요청하기 위함
-        tokenManagementService.addRefreshTokenToCookie(request, response, refreshToken);
+        tokenManagementService.addRefreshTokenToCookie(request, response, refreshToken, loginRequest.getRememberMe());
 
         // Access Token 생성
         String accessToken = tokenProvider.generateToken(user, TokenManagementService.ACCESS_TOKEN_DURATION);
@@ -115,43 +127,82 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
         // 응답 헤더에 액세스 토큰 추가
         headers.set("Authorization", "Bearer " + accessToken);
 
-        //  세션이나 쿠키에 불필요한 데이터가 남아 있지 않도록 하여 보안을 강화함.
+        //  인증 실패와 관련된 정보를 세션에서 제거. 즉 다음에 재로그인할때 만약 이전 인증 실패 정보가 남아있다면 이전 인증 실패 정보가 남아있지 않도록 함.
         super.clearAuthenticationAttributes(request);
 
-        return ResponseEntity.ok().headers(headers).body("Login Success");
+        return ResponseEntity.ok().headers(headers).body(new SuccessResponse<>("로그인에 성공하였습니다."));
 
     }
 
+
     // Custom logout 로직을 구현한 경우 시큐리티에서 제공하는 logout을 사용하지 않는다.
     @PostMapping("/api/users/logout")
+    @Transactional // deleteRefreshToken DB작업 있어서 트랜잭션 추가해야함
     public ResponseEntity<String> logout(HttpServletRequest request, HttpServletResponse response) {
 
-        System.out.println("실행 로그아웃");
-        CookieUtil.deleteCookie(request, response, "refresh_token");
-        CookieUtil.deleteCookie(request, response, "access_token");
-
         String authorizationHeader = request.getHeader("Authorization");
+
+        if (authorizationHeader == null || !authorizationHeader.startsWith("Bearer ")) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("로그인 상태가 아닙니다.");
+        }
+
         // "Bearer " 이후의 토큰 값만 추출
         String token = authorizationHeader.substring(7);
 
-        try {
+        CookieUtil.deleteCookie(request, response, "refresh_token");
+        CookieUtil.deleteCookie(request, response, "access_token");
 
-            Long userId = tokenProvider.getUserId(token); // tokenProvider.gerUserId()에서 내부적으로
-            // 만료된 토큰인지 유효성 검사를 함. 이때 만료된 토큰이면 아래 ExpiredJwtException Catch문으로 넘어간다.
-            refreshTokenService.deleteRefreshToken(userId);
-            return ResponseEntity.ok("Successfully logged out.");
+
+
+        try {
+            // tokenProvider.gerUserId()에서 내부적으로 만료된 토큰인지 유효성 검사를 함. 이때 만료된 토큰이면 아래 ExpiredJwtException Catch문으로 넘어간다.
+            String email = tokenProvider.getEmail(token);
+
+            redisTemplate.delete(TokenManagementService.RT_PREFIX + email);
+
+            return ResponseEntity.ok("로그아웃에 성공하였습니다.");
 
         } catch (ExpiredJwtException e) {
             // 만료된 토큰일 때도 userId를 추출 가능 (ExpiredJwtException을 통해 Claims에 접근) 즉
             // ExpiredJwtException을 통해 만료된 토큰에 있는 Claims에 접근한다.
-            Long userId = e.getClaims().get("id", Long.class);
-            refreshTokenService.deleteRefreshToken(userId);
-            return ResponseEntity.ok("Successfully logged out with expired token.");
+            String email = tokenProvider.getEmail(token);
+            redisTemplate.delete(TokenManagementService.RT_PREFIX + email);
+            return ResponseEntity.ok("로그아웃에 성공하였습니다.");
 
         } catch (Exception e) {
+            log.error("Error deleting refresh token for userId: ", e);
             // 유효하지 않은 토큰(서명이 잘못되거나 변조된 경우 등 즉 비정상적인 토큰일 경우)이면 거부 한다.
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Invalid token.");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("유효하지 않은 토큰입니다.");
         }
+
+    }
+//
+//    @GetMapping("/api/{userIdentifier}/availability")
+//    public ResponseEntity<ApiResponse> checkUserExists(@PathVariable("userIdentifier") String userIdentifier) {
+//
+//        log.info("userIdentifier: " + userIdentifier);
+//
+//        if (userService.existsByUserIdentifier(userIdentifier)) {
+//            return ResponseEntity.ok()
+//                    .body(new SuccessResponse<>("User exists"));
+//        }
+//
+//        return ResponseEntity
+//                .status(HttpStatus.NOT_FOUND)
+//                .body(new ErrorResponse("User not found.", 404));
+//    }
+
+    @GetMapping("/api/check/blog-id/exists/{blogId}")
+    public ResponseEntity<ApiResponse> checkExistsBlogId(@PathVariable("blogId") String blogId) {
+
+        if (userService.isExistsBlogId(blogId)) {
+            return ResponseEntity.ok()
+                    .body(new SuccessResponse<>("User exists"));
+        }
+
+        return ResponseEntity
+                .status(HttpStatus.NOT_FOUND)
+                .body(new ErrorResponse("User not found.", 404));
     }
 
     @Operation(summary = "블로그 ID 중복 확인")
@@ -173,11 +224,10 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
             )
     })
     @Parameter(name = "blogId", description = "확인할 블로그 ID", required = true)
-    @GetMapping("/api/check/blogId/{blogId}")
-    public ResponseEntity<ApiResponse> checkBlogId(@PathVariable("blogId") String blogId) {
-
-
-        DuplicateCheckResponse response = userService.existsByBlogId(blogId);
+    @GetMapping("/api/check/blog-id/duplicate/{blogId}")
+    public ResponseEntity<ApiResponse> checkDuplicateBlogId(@PathVariable("blogId") String blogId) {
+        
+        DuplicateCheckResponse response = userService.isDuplicateBlogId(blogId);
 
         return checkDuplicate(response);
     }
@@ -201,10 +251,10 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
     )
     })
     @Parameter(name = "email", description = "확인할 이메일", required = true)
-    @GetMapping("/api/check/email/{email}")
-    public ResponseEntity<ApiResponse> checkEmail(@PathVariable("email") String email) {
+    @GetMapping("/api/check/email/duplicate/{email}")
+    public ResponseEntity<ApiResponse> checkDuplicateEmail(@PathVariable("email") String email) {
 
-        DuplicateCheckResponse response = userService.existsByEmail(email);
+        DuplicateCheckResponse response = userService.isDuplicateEmail(email);
 
        return checkDuplicate(response);
 
@@ -229,11 +279,11 @@ public class UserApiController extends SimpleUrlAuthenticationSuccessHandler {
             )
     })
     @Parameter(name = "username", description = "확인할 사용자명", required = true)
-    @GetMapping("/api/check/username/{username}")
-    public ResponseEntity<ApiResponse> checkUserName(@PathVariable("username") String username) {
+    @GetMapping("/api/check/username/duplicate/{username}")
+    public ResponseEntity<ApiResponse> checkDuplicateUsername(@PathVariable("username") String username) {
 
 
-        DuplicateCheckResponse response = userService.existsByUserName(username);
+        DuplicateCheckResponse response = userService.isDuplicateUsername(username);
 
         return checkDuplicate(response);
     }
